SinGAN/requirements.txt
# the project was written to be compatible with python 3.6
# prepared for conda, change pytorch to torch for pip
numpy
matplotlib
pytorch
torchvision
scipy==1.3.3
tqdm
